CustemUserDetails added

+++ src/main/java/com/projects/library/repository/UserEntityRepository.java
@@ -2,8 +2,9 @@ package com.projects.library.repository;
 
 import com.projects.library.model.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.security.core.Authentication;
 
-public interface UserEntityRepository extends JpaRepository<UserEntity, Integer> {
+public interface UserEntityRepository extends JpaRepository<UserEntity, Long> {
     UserEntity findByUsername(String username);
     UserEntity findByEmail(String email);
 }
